Rework db seed template
